Moved comments into component folder
